Solution to Assignment 4

+++ kern/include/proc.h
@@ -49,7 +49,7 @@ struct file_info {
 	struct vnode * file;
 	struct lock *fd_lock;
 	int status_flag;
-	size_t offset;
+	off_t offset;
 	int ref_count;
 };
